Everyting for Electron ID from Tongguang and Ahmad

+++ UFHZZ4LAna/src/UFHZZ4LAna.cc
@@ -1505,8 +1505,8 @@ UFHZZ4LAna::findHiggsCandidate(std::vector< pat::Muon > &selectedMuons, std::vec
                 }
 
 
-                double phopt_i=0.0;
-                double phodr_i=9999.9;
+                float phopt_i=0.0;
+                float phodr_i=9999.9;
 
                 Za = tmpZa;//hualin
                 Zb = tmpZb;//hualin
